moved PrivateRoute.js and AuthContext.js into auth directory
